move allByKeys to shared/api
