+++ shared/api.js
@@ -3,7 +3,7 @@
 const getFromApi = async (api, url) => {
   const { data } = await api.get(url);
   return data;
-}
+};
 
 const getSingleItem = async (api, url, key) => {
   // The api is case insensitive when getting by url/login
@@ -13,12 +13,12 @@ const getSingleItem = async (api, url, key) => {
   if (data[key]) {
     return data[key];
   }
-  const realKey = Object.keys(data).find(dataKey => dataKey.toLowerCase() === key.toLowerCase());
+  const realKey = Object.keys(data).find((dataKey) => dataKey.toLowerCase() === key.toLowerCase());
   if (realKey) {
     return data[realKey];
   }
   return null;
-}
+};
 
 const getAllPages = async (api, url) => {
   let hasMore = true;
@@ -32,8 +32,20 @@ const getAllPages = async (api, url) => {
   return results;
 };
 
+// like Promise.all but with an object instead of an array, e.g.
+// `let { user, projects } = await allByKeys({ user: getUser(id), projects: getProjects(id) })`
+const allByKeys = async (objOfPromises) => {
+  const keys = Object.keys(objOfPromises);
+  const values = await Promise.all(Object.values(objOfPromises));
+  return keys.reduce((result, key, i) => {
+    result[key] = values[i];
+    return result;
+  }, {});
+};
+
 module.exports = {
   getFromApi,
   getSingleItem,
   getAllPages,
+  allByKeys,
 };
